main/requirements.txt
ultralytics
cvzoneS